Remove unnecessary .idea configuration files and update gitignore
Removed several .idea configuration files to reduce clutter in the repository. These files were specific to the IDE and should not have been tracked in the first place. Updated the .gitignore to ignore these files in the future. Minor adjustments were also made to the ".github/workflows/pylint.yml" file.