Add database operations (#12)
* Implement `DatabaseManager` to manage database operations and update CLI interface accordingly.

This commit adds a new file `DatabaseManager.py` that provides functions for database management like creation, deletion, and backup. It is implemented as a class that represents a database with methods to handle the aforementioned operations. The `EnvManager.py` file is altered to permit reading of a default value for keys. Many test files are also updated to reflect these changes.

* Update `DatabaseManager` class to support password and port.

Implemented password and port parameters in the `DatabaseManager` class. Also added a `connect` method to establish connections. Updated the tests and main application accordingly to support these changes.

* Remove unnecessary .idea configuration files and update gitignore

Removed several .idea configuration files to reduce clutter in the repository. These files were specific to the IDE and should not have been tracked in the first place. Updated the .gitignore to ignore these files in the future. Minor adjustments were also made to the ".github/workflows/pylint.yml" file.

* Refactor DatabaseManager and update related tests

The DatabaseManager class has been refactored to encapsulate add, remove user and SQL execution operations. Previously, these operations were executed in ad-hoc manner. Additionally, connecting to the database has been made private, while test cases have been renamed and moved to reflect the application structure.

* Refactor test suite structure and locations

To align with the refactoring of the DatabaseManager class, the associated test cases have been renamed and moved. This change makes the test file locations more accurately reflect the structure of the application, promoting easier navigation and better organization within the testing suite.

* Update Python version in GitHub workflows

The Python version used in the GitHub workflows has been updated from 3.10 to 3.11. This will ensure that the project is tested against the latest Python version, keeping the software up-to-date and